Moved feed modules tests from subfolders into main folders.
